requirements.txt
openai==1.3.9
opencv-python==4.8.1.78
selenium==4.16.0
selenium-wire==5.1.0
webdriver-manager==4.0.1
environs==9.5.0
beautifulsoup4==4.12.2
typer[all]==0.9.0
Pillow==10.1.0
faiss-cpu==1.7.4
tiktoken==0.5.2
numpy==1.26.2
sentence-transformers==2.2.2